requirements.txt
asgiref==3.2.10
attrs==20.2.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
cymem==2.0.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.10
django-import-export==2.4.0
django-mssql-backend==2.8.1
docx2txt==0.8
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
joblib==0.16.0
jsonschema==3.2.0
MarkupPy==1.14
murmurhash==1.0.2
nltk==3.5
numpy==1.19.2
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.2
pdfminer.six==20200726
plac==1.1.3
preshed==3.0.2
pycparser==2.20
pycryptodome==3.9.8
pyodbc==4.0.30
pyresparser==1.0.6
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
tablib==2.0.0
thinc==7.4.1
tqdm==4.49.0
urllib3==1.25.10
wasabi==0.8.0
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
